+++ lib/assertion/dsl/attributes.rb
@@ -39,7 +39,7 @@ module Assertion
 
       def __add_attribute__(name)
         __check_attribute__(name)
-        define_method(name) { attributes[name] }
+        define_method(name) { attributes.fetch(name) }
         attributes << name.to_sym
       end